disable special drawings of the titlebar for now
the online/offline "shadow" is not used in all viewcontrollers
and is not always in sync with the current state.
also onboarding is a bit weird.
so, if we really want to use such an indiactor,
this would require some more investigation.
also, with the commands commented out by this commit,
for some reasons, the opaque-look-n-feel is not ios-ish
and has additional problems in dark-mode.