move contact-request aka mailbox out of focus
the idea is that contact-requests pop up directly in the chatlist.
so, the contact-request-view is only needed
when the user discards a request and decides to come back to that later.
in a typical usage, there is often no need to open the contact requests,
therefore, a whole tab, at eye level of "chats" is overrated.

on android, contact-requests are used in the same way,
being not directly in the eyes of the user.

there are still some things to do on contact requests,
eg. they are still a bit ugly and not always refreshed in the chatlist
and the options show_emails=all should display all incoming mails
as contact-requests directly in the chatlist.