streamline port selection
the old "featured" view controller brough more problems than it solved,
eg. ports were shown as "0" instead of empty.
also it lacked an "automatic" option.
also it was a bit unclear when the controller
was left if the port was saved or not
(it was not)

instead of fixing all that, just get rid of the controller at all.
a simple input field for the port selection is just fine,
also in-line with android and desktop.