use textLabel for selected autodel-value
the field-name goes to the headline.
this might not be 100% the intended way of using the cell,
however this results in a much clearer view, without truncated values or labels,
especially on smaller screens.
and, in an radio-view, the textLabel is used for values as well.

an alternative might be UITableViewCell with "style: .subtitle",
however, this brings the value out of focus -
and is not the intended way as well.