to go chat on mute/disappearing changes
this pr pops directly to the chat-view
when changing mute/disappearing settings,
the profile-view is skipped.

the profile is partly used as a menu,
this change underlines that.

also, for the disappearing settings,
a message is generated and sent,
this way, this is directly visible to the user.
otherwise, this fact will be a bit hidden.