adapt to new dc_configure() error handling
errors that should be shown to the user
come from DC_EVENT_CONFIGURE_PROGRESS(data1=0) as data2
since https://github.com/deltachat/deltachat-core-rust/pull/1905;
there is no need to show the captured error any longer
(which was unreliable and hard to handle from core site).

there is some capturing-code left as needed for other parts,
but it is not used for configuration anymore.