rely on the fact that there is always a quote text
core guarantees that there is always a quote text if there is a quote,
in QuotePreview::configure(), rely on that,
using a summary fallback is not needed.

moreover, in DraftModel::setQuote(), we use the core to set the quote summary.
this closes #996

(nb: quote summaries were alrady fine before when a
chat with a draft-quote was left and entered again,
this was because of the implicit draft creation that is now
done explicit in DraftModel::setQuote())