disable 'select more' if the chat cannot sent
this is a quick fix and affects the device chat and contact requests only,
reason is that 'select more' needs the messageInputBar to work,
and the messageInputBar is not available if one cannot send in the chat.

we might think that over when we have more chats that cannot send
(upcoming mailing lists maybe).