call new dc_accounts_maybe_network_lost()
when calling dc_accounts_maybe_network_lost() on lost network,
we get the corresponding connectivity-changed events from the core,
delayed by 10~20 seconds, for whatever reason,
maybe cores tries some things ...
but that seems to be okay.

when networks becomes available,
the update is instant.

the "crash" on network lost, btw seems not to be a crash,
it is more that the connection to the debugger ist lost,
maybe because of missing network.

without debugger,
nothing crahes.