add 'connectivity changed' observers in time
'connectivity changed' observers must be added before
the UI is inititalized - otherwise,
we may miss an event and UI is stuck at a wrong state.

in general, that means to move adding observers to viewWillAppear -
which seems also be more balanced as they're removed on viewDidDisappear -
so both in situations the view is not visible.