show unread-message-badge for chat requests
chat requests are notified since core61
and may also be added to the app-icon-unread-badge by the system therefore.

to not confusing users,
which chat is unread when tapping the app icon with the unread-badge,
we should show the unread-message-badge also for chat requests.

previously it was hidden with the reasoning
that chat requests are not notified,
https://github.com/deltachat/deltachat-ios/issues/1306
this has changed since then, by strong user requests :)
cmp https://github.com/deltachat/deltachat-core-rust/pull/2690