stop ChatViewController timings
on my iphone7, 3gig db:

- without the recent background_image pr,
  80ms are needed to set up views,
  _plus_ time needed for getChatMsgs(),
  so, overall, 0.2 to 1.0 s

- _with: the recent background_image pr,
  100ms are needed to set up views.

so, all in all, backround_image pr add 20 ms,
on a quite old phone, mind you.

i think, we should streamline the measurement,
maybe removing all but the last line,
so that we can get a feeling about timings
on different devices with different wallpapers.