do not send webxdc updates if lastSerial is unset
that speeds up webxdc starts as the updates are not fired "for nothing"
if the handler ist not set yet.

moreover, that removes a theoretical race
if an update happens between `update_listener = cb`
and `webkit.messageHandlers.setUpdateListener.postMessage`.

(in practise, both, dcNotificationWebxdcUpdate() and userContentController()
are called on the main thread so there should have been not races in practise,
however, for the future, this may change)