call performFetch() also from didFinishLaunchingWithOptions with launchOptions?[.remoteNotification] set
while most times, this also results in didReceiveRemoteNotification being called
this is not always the case.
not sure at all, what these two entry points are needed for,
you do not find many information about that,
so it is good to be prepared for everything.
also, reading through stackoverflow, things have changed over time and iOS versions ...