signal when connectivity is "connected" again
this finishes the background-fetch much faster,
instead of always 11 seconds, things finish in less than 2 seconds usually
and take longer only when here is really something to download.

fetchSemaphore is signalled when we receive the DC_CONNECTIVITY_CHANGED event
with connectivity set to DC_CONNECTIVITY_CONNECTED - we have called maybeNetwork()
just before which sets connectivity to CONNECTING,
we know pretty sure that one round is done.

we check that on the selected context only,
that is maybe not perfect, but probably good enough
(other accounts are checked as well, but may be stopped too soon)