double max. fetching time to 20 seconds
before, we were always using exact 10 seconds to not eat too much battery.
as we now terminate much sooner, usually in less than 2 seconds,
we can also allow to take more time as needed.
(in total, we can spend 30 seconds in the background fetch,
the remaining 10 seconds are saved for shutdown and as headroom)