show 'back button' in preview again
`QLPreviewController` seems not to show a 'back' button
if `present()`ed inside a `UINavigationController` -
in the past, there was a hack that adds the button manually in this case.

this hack stopped working with ios16 -
resulting having no way to leave preview.

this pr removes the hack
and stops using `present()` and switches to `pushViewController()` -
this seems to be the natural way here and shows the proper back buttons.

moreover, the pr removes the sometimes used wrapper ("betterPreviewController"),
not sure what was the reasinging for that, probably needed for some fullscreen-mode
however, pushing a `UINavigationController` to a `UINavigationController` is not allowed,
so it just cannot be used in this form anymore -
and we're "fullscreen" as usual anyway.

closes #1696