synchronize color and size more to android and other online indicators (#1707)
some easy to target layout points:

- with the "connection view",
  we already have a green indicating sth. as "online", reuse that
- the size is a bit bigger, more matching to android
  and also more visible in the title bar (where the avatar is smaller)

maybe we can get online indicator also to the top-right corner,
maybe above the checkmark that could move a little bit to the left (always),
i was playing around with this and other ideas,
but that is maybe sth. for a subsequent pr. easy things first :)