move "recently seen dot" lower right
this moves the "recently seen dot" to the corner where it is shown
also for many other apps: to the lower right corner.

as the lower right corner was already occupied by the "verified checkmark",
that one is moved to the upper right corner, with a little bit of padding to the top.
i played around with different positions, that one looked best.

also the combination of both indicators look fine now.

it is anyway a bit unclear what happens to the "verified" indicator
for contacts, it may go away completely and being replaced by
a line as "Verified by NAME", so it is good enough for now :)