accessibility improvements (#1738)
* avoid unaccessible avatars in group chats (appearing as empty entries in VoiceOver), add custom accessibility actions to handle common message actions more easily

* Trim down accessibility/navigation items in AudioRecorderController if VoiceOver is enabled. During recording the screen reader is necessarily muted, so that VoiceOver users cannot orientate themselves auditorily and thus not navigate safely.

* revert accessibility changes that require changing the UI dynamically

* use the normal context menu for accessibility

i thought quite a while over that and played around with VoiceOver,
however, finally came to the conclusion that it is in our situtation,
with very few developmenu resources,
better to leave the accessibility options as close to the normal layout as
possible.

we should spend the resources for accessibility first of all
to provide (1) labels and help (as we do)
and (2) to tweak layouts if things really get unusable otherwise (as we do eg. for voice messages).
however, in (3) other situations, go for the defaults if things are accessible already by that [^1]

we would otherwise risk to getting worse at (1) and (2) or at overall app development
when spending too much resources at (3) (eg. this effort would need to be continued for consistency)

[^1]: the context menu is available quite simple by
  double-tab-hold, this is probably know from other apps already.
  https://www.applevis.com/forum/ios-ipados/how-do-you-perform-long-press-action-using-voice-over

* re-optimize getting sender name

Co-authored-by: B. Petersen <r10s@b44t.com>