update title unconditionally in CHAT_MODIFIED
this will lead to double updates as
the title is updated in viewWillAppear as well
if viewControllers atop are removed.
however, this is probably only a minor issue
and the correct way seems to be to alter viewWillAppear
as doing fewer things there and having the ViewController "ready" before coming to view
feels snappier anyway.