swap 'Gallery' and 'Files' tab
that way, two taps in different areas are required,
making it harder to show maybe private images accidentally.

these two clicks is also how android works since some time in Delta Chat -
and even longer in Delta Lab - without complains in that direction.

all this is meant only as a minor precaution against accidental taps,
not as something that helps if someone plays around with unlocked
phone for minutes.
for that, we have ephemeral messages.