use translucent qr-code title bar
the old, totally transparent title bar led to more or less
unreadable tabs, depending on what is shown to the camera.

after figuring out the translucent thingie at #1843,
this is a nice improvement for the other view using a tabbar in the title.

moreover, this pr adds an animation for both tabs,
things really seem to be fast enough,
also, the system seem to do UIViewController creation in parallel anyways.