no delete/pin/archive swipe for messasge results
even though message results have a chatId,
delete is ambiguous as ppl may think,
a single message gets deleted, not a whole chat;
better leave out this option therefore.

for pin/archive, this is discussion-worthy and not that clear,
however, for now, leave out these options as well
to be in sync with desktop and android here.