allow pinning/archiving of message search results
the message search results are shown as chat items,
including indicators about pinning, archived etc.
there are no good reasons that you cannot eg. unpin a
chat found by message search;
it is pretty clear that this is about chats and not messages -
and if not, it is easily undoable.

targets #1869