Streamline group creation (#1953)
* remove 'create verified group' from 'new chat'

* remove qr-invite option from create-group

* remove isVerified branches from create-group

* ask whether to create verified group if all members are verified

* the groupId is never present before creation now

* cleanup

* simplify member list loading

* remove weird comment

* use .actionSheet for verified/unverified question

this fits better as it is a result of a direct user action;
also looks better with the additional 'Cancel' button
(otherwise user may get stuck, which does not feel good)

* more concrete 'Create' buttons

us term 'Create' instead of 'Done';
this is esp. better in case there is a 'Done' already when adding members.

* simplify chat object creation, remove dead code

* show qr code when tapping an unverified member in a verified group

* remove create (un)verified question

it is hard to answer,
better go with defaults here until we get
a real usecase where we would need this question.