let logger accept strings only
the few occurances with other objects look to me more like accidents,
more that one is really knowing what one is doing.

it is better to convert to string on caller side,
unless we really have a use of letting loggers accept randome objects.
this also makes it easier to switch to os_log() or whatever