prescale images to fit into extension restrictions
extensions must not use more than 120mb of memory,
this is easily exhausted by core scaling down images.

using iOS function for scaling use less memory (or the memory is not counted)
and avoids this error.

although we accept various formats of images,
in practise i've seen them passing to the extension only as URL,
if we see that for other formats as well, we can optimize the other branches,
currently, that seems not to be needed.