focus on name and avatar in guaranteed-e2ee chats
this PR hides the email address from the chat's title bar
for guaranteeedd e2ee one-to-one chats.

this affects chatmail contacts and delta-to-delta contacts mainly,
for classic email usage, things stay the same mostly.

to be clear: the email address will stay in profile (one tap away)
and there are no ideas to change that.

for reasons and discussion,
see https://github.com/deltachat/deltachat-android/pull/2916 ,
to sum up:
- protects against "over the shoulder surveillance",
  a real-world issue reported
- the address is not meaningful for chatmail.
  featuring chatmail is importing for future users,
  esp. <30yo ones considering email a
  "thing of the past that can't possible bring any good" :)
- act similar to other messengers, not showing the "handle" for known contacts
- focus are shift to name and avatar, not clutterung UI,
  making guaranteed e2ee more outstanding
- impersonation won't become worse,
  you could always use trustworthy looking but wrong addresses.
  removing the address may even improve here
  as other aspects (chat history etc.) may get more into focus.
- prepare for AEAP and the idea of multilple addresses per account