do not add "Changelog" for new accounts
on new installations or accounts,
the "What's new?" messages only clutter the device messages
with unimportant information
(either the user is new to the app, so everyrhing is new and nothing is changed,
or an existing user has created a new account, they have seen the changelog anyways)

in other words,
the changelog will be added only for updates.

when reading https://c.delta.chat/classdc__context__t.html#a1a2aad98bd23c1d21ee42374e241f389 , this was always the gist,
however, somehow forgotten or removed on the way :)

technically, it was easier to track the creation in an array,
as the user flow from creation to finally addind the device message
is a bit too complicated to pass flags around.