allow to retract any reaction
currently, random reactions (as added with #2087)
cannot be retracted easily.

this PR allows retraction by just selecting the emoji again -
similar to the predefined reactions.

(showing the selection in the emoji selector is out of scope of this PR
and probably not worth the effort,
at least not now where there are much larger fishes to get)