make use of notifications filtering entitlement
without the entitlement `com.apple.developer.usernotifications.filtering`,
in case an notification is irrelevant,
we need to adding it anyways as it is not possible to _not_ show a notification.
(to mitigate that situation, we do not play a sound, do not light the display,
and remove the notification asap - but it may stay for tens of minutes)

with the entitlement,
we can just not show a notification.