always show the selected reaction in menu
if a reaction was selected,
show that instead of "...".

this is what whatsapp/signal are doing - as well as
deltachat android since
https://github.com/deltachat/deltachat-android/pull/2979

deltachat desktop has more room and shows "..." and selection the same time.

little drawback is that changing a reaction requires two taps,
but that seems fine in practise and is outweighted by advantages as:
- always seeing what was selected
- make it clearer, there is only one reaction
- make it clearer how to retract ones reaction
- retracting is always only one tap
- consistency