Add buttons and stackview-magic
We want those two buttons to be at the bottom of the screen, but they should come up when the keyboard appears.
This is the reason why there are two `UIView`s called spacers. Combined with constraints and StackView, we update the layout when the keyboard appears or disappears.
When users scroll, the keyboard disappears