emphasize 'Existing Password'
make clearer, this is not a password the user can define here,
esp. when input is started or 'Password & Account' is opened later,
this information was hidden.

we found out by complains / user testing in the past
that the term 'Existing' is needed here,
cmp https://github.com/deltachat/deltachat-android/pull/1256 .

in case of chatmail creating accounts
by using new passwords for non-existing addreses,
this is a special, not recommended case anyways -
and you still cannot change passwords eg. on editing account setings.