add profile on tap/scan QR code unconditionally
when tap/scan QR code after hitting "Add Profile",
things are fine, a new profile was added explictly.

however, when tap/scan QR code when a profile is just opened
normally (in chatlist or chat),
we need to add an account and not use the existing one.

(we cannot just always create an account when entering the viewController
as we want things as name and avatar being persisted when switching
shortly to the browser to select another chatmail server)

counterpart of https://github.com/deltachat/deltachat-android/pull/3112