show context menu on long-tapping proxies
this PR shows a context menu with the options 'Share' and 'Delete'
when a proxy item is long-tapped.

historicaly, swiping a row left/right
are the default way where ppl expect actions,
esp. as context menus only were added on iOS 13.

however, it happend to me regulary, that i try to long tap -
and i have seen that also by others that are using the app,
seems long-tap is quite arrived on iOS,
so, let's just support both ways
(as we stop supporting iOS 12, these menus are much more accessible for devs :)