defer attach menu creation
this PR creates the attach menu only if needed;
this was also the case in the past, when we were using an alert instead of modern menu.

it is only a little bit faster currently (2~3ms vs <0.5ms on iphone7),
however, as this also may be used as a pattern and as slower conditions may be added in the future,
maybe still worth going for it.
also to exclude that part from being slow.