contact profile: hide unavailable media/location
if there is no chat with a contact,
in the contact's profile,
hide the superfluous media and location options -
similar to eg. shared chats and buttons as clear/delete chat.

this unclutters the profile,
puting things as "send message" more into focus.
and it looks better,
there are otherwise a bit too many grey tones in the upper area :)