fix broken 'Add Second Device' QR code
core 1.155.2 accidentally broke compatibility
to add-second-device QR codes of previous versions;

this is esp. bad as ppl will have different versions for some days at least,
weakening overall UX, esp. of first-time-users
that may come to delta because of praised, seamless multidevice ... :)

this PR is a little experiment, if it is really only the minor JSON change,
or sth deeer.

seems, these two repacements fix the issue,
allowing us to release much mess stressful,
while users being happy.

if we think, this is fine,
we should do that better in core, however.

the hack can be removed 3 releases later or so.