deduplicate files on sending
when https://github.com/deltachat/deltachat-core-rust/pull/6558 gets merged,
the old setFile() function is no longer available.

instead, setFileAndDeduplicate() should be used,
that renamed the file if it is already in the blob directory.

i checked common path logic-wise,
they are not passing blob-dir or do not rely on the file to exist afterwards.

still, general things should be tested after merging,
so we should merge this in soon :)