faster opening of profile and connectivity
before, to get some immediate visual feedback,
we disabled the title and then pushed the view controller on the next main loop.
this was kind of okay when profile opening took a lot of time.

meanwhile, profile opening is much faster,
so that this additional look becomes the bottleneck.

this PR changes to push the view controller immediately (as usual elsewhere) -
for immediate feedback, CATransaction.flush() is used.

same for the connectivity,
where the same pattern was used.