add shortcut to start in-chat search
in-chat search is started usually from the profile menu,
which is straight-forward,
but requires some taps (tap title, tap menu, tap search)

this adds an shortcut to open the search by long-tap title.

this also replaces the extra search icon for the saved messages -
for consistency, but also as we have with the 'app & media' another button there.