don't show ... in subtitle on most left groups
checking for self-in-group was introduced at #2895
to avoid showing misleading "1 member" during securejoin,
when the member list is not known.

this has the disadvantage that the ... subtitle is also shown for left groups.

at https://github.com/deltachat/deltachat-android/pull/4017
we improved the condition for showing member count
slightly to "member-count > 1 || self-in-group",
see there for more reasoning and discussion

this is still a bit hacky,
but that's it :)